refactor: Remove shelljs dependency and instead use `node:child_process`
The `shelljs` package is supposed to bring some unix
commands as a cross-os compatible API (e.g. cd, mkdir, etc.).
But we are not using any of those and only run `exec`.

Node.js already provides an API for executing the
commands we want, so we can get rid of this dependency.