feat(configurable-boilerplate): Interactively ask the user all needed variables
I decided on using the `@inquirer/prompts` package and
used it for all interactive questions.
In addition to the boilerplate question, I also added the
project-name and output-dir as questions, which makes
passing these values in the CLI optional in the future.

I think the nice interface makes the dependency worth it
and its dependency tree looks simple enough, that it should
hopefully not cause any upgrade problems in the future

#6