refactor(configurable-boilerplate): Extract CLI argument parsing into a class
This hopefully helps in making the `index.js` file easier to
read and understand.
I also made sure that providing an empty string as an
argument still uses the fallback.
No fancy validation of any kind, if you want to provide a
white space as the (default) project name: You are on your own.

#6