fix(configurable-boilerplate): Allow using the `--loglevel` in any position
The previous implementation relied on the positional
arguments (projectName and outputDir) to be set for
`--loglevel` to work properly.
Looking at the implementation in the main branch, I do
not understand how it would work there probably.
Maybe the currently executed code is different from it.

This also makes the parsing way nicer that before, as we
can rely on the Node.js function instead of doing manually.

#6